Decompose 'resizeImage'
Decompose line by line the resize image method, for educational purposes.
We first make an object ImageIcon with the path specified in the params, then we change it to a Image object, so we can then get an instance resized and finally convert it back to an ImageIcon object, the one that the method will return.